projects/requirements-base.txt
werkzeug.security
webbrowser
views
utils
urllib
typing
time
textwrap
test
tabulate
sys
subprocess
strings
string
streamlit
signal
seaborn
requests
re
random
pyperclip
pypandoc
psutil
praw
plotly.graph_objs
plotly.graph_objects
plotly.express
plotly
parallel
panflute
pandas
os
openpyxl
numpy
multiprocessing
matplotlib.pyplot
logging
lib.helpers
json
io
google.oauth2
google.generativeai
google.cloud.documentai_v1
google.cloud
google.auth.transport.requests
google.api_core.exceptions
google.api_core.client_options
googleapiclient.http
googleapiclient.discovery
functools
flask
enum
dotenv
docx
datetime
opencv-python
contextlib
concurrent
bs4
base64
auth
argparse
